20161017 更新了部分代码，统计食物摄入， 统计营养摄入还没做完； 数据库部分属性更新， 补充了属性列，更改了属性列的type。

+++ NutritionInvestigation/NutritionInvestigation/FoodClass.cs
@@ -26,7 +26,7 @@ namespace NutritionInvestigation
         public string ParentName { get; set; }
         public Nullable<long> Percent { get; set; }
         public Nullable<long> Edible { get; set; }
-        public Nullable<long> RecordMode { get; set; }
+        public Nullable<int> RecordMode { get; set; }
         public string TypicalFood { get; set; }
         public string FoodIndex { get; set; }
         public string FoodUnit { get; set; }
